pyproject.toml
[tool.poetry]
name = "django-mongodb-proj"
version = "0.1.0"
description = ""
authors = ["Jason Xue <jasonxue1228@gmail.com>"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.6"
djangorestframework = "^3.12.4"
djongo = "^1.3.4"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
